requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0

# Vision
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
opencv-python>=4.5.0
albumentations>=1.3.0

# NLP
transformers>=4.25.0
sentence-transformers>=2.2.0
tokenizers>=0.13.0

pycocotools>=2.0.6
datasets>=2.8.0
pandas>=1.5.0
wandb>=0.13.0
tqdm>=4.64.0
omegaconf>=2.3.0
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.11.0
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=6.5.0
pyyaml>=6.0
rich>=12.0.0
loguru>=0.6.0
